pyproject.toml
[project]
name = "explainable-melanoma"
version = "0.1.0"
description = "Explainable Melanoma Detection"
authors = [
    {name = "Sukhrob Ilyosbekov", email = "suxrobgm@gmail.com"},
]
readme = "README.md"
license = {text = "MIT"}
requires-python = "<3.15,>=3.14"

dependencies = [
    "Pillow>=12.0.0",
    "tqdm>=4.67.1",
    "rich>=14.2.0",
    "torch>=2.9.0",
    "torchvision>=0.24.0",
    "numpy>=2.2.6",
    "pandas>=2.3.3",
    "PyYAML>=6.0.3",
    "scikit-learn>=1.7.0",
    "matplotlib>=3.10.0",
    "seaborn>=0.13.0",
    "opencv-python>=4.12.0",
    "grad-cam>=1.5.5",
]

[dependency-groups]
dev = [
    "black>=25.11.0",
    "isort>=7.0.0",
]

[tool.pdm.scripts]
train = "python scripts/train.py --config config.yaml"
eval = "python scripts/eval.py --config config.yaml --checkpoint checkpoints/best_model.pth"
infer = "python scripts/infer.py --config config.yaml --checkpoint checkpoints/best_model.pth --input-dir data/sample_images"
generate-concepts = "python scripts/generate_concepts.py --config config.yaml --output data/concepts"
train-fastcav = "python scripts/train_fastcav.py --config config.yaml --checkpoint checkpoints/best_model.pth"


[tool.pdm]
distribution = true

[tool.pdm.build]
package-dir = "src"
includes = ["melanomanet"]

[tool.pdm.resolution]
respect-source-order = true

[[tool.pdm.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu130"
verify_ssl = true

[[tool.pdm.source]]
name = "pypi"
url = "https://pypi.org/simple"
verify_ssl = true

[tool.isort]
profile = "black"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
